+++ tests/library/browsercontext-add-cookies.spec.ts
@@ -357,75 +357,6 @@ it('should set cookies for a frame', async ({ context, page, server }) => {
   expect(await page.frames()[1].evaluate('document.cookie')).toBe('frame-cookie=value');
 });
 
-it('should(not) block third party cookies', async ({ context, page, server, browserName, allowsThirdParty }) => {
-  await page.goto(server.EMPTY_PAGE);
-  await page.evaluate(src => {
-    let fulfill!: (value: unknown) => void;
-    const promise = new Promise(x => fulfill = x);
-    const iframe = document.createElement('iframe');
-    document.body.appendChild(iframe);
-    iframe.onload = fulfill;
-    iframe.src = src;
-    return promise;
-  }, server.CROSS_PROCESS_PREFIX + '/grid.html');
-  await page.frames()[1].evaluate(`document.cookie = 'username=John Doe'`);
-  await page.waitForTimeout(2000);
-  const cookies = await context.cookies(server.CROSS_PROCESS_PREFIX + '/grid.html');
-  if (allowsThirdParty) {
-    expect(cookies).toEqual([
-      {
-        'domain': '127.0.0.1',
-        'expires': -1,
-        'httpOnly': false,
-        'name': 'username',
-        'path': '/',
-        'sameSite': 'None',
-        'secure': false,
-        'value': 'John Doe'
-      }
-    ]);
-  } else {
-    expect(cookies).toEqual([]);
-  }
-});
-
-it('should not block third party SameSite=None cookies', async ({ contextFactory, httpsServer, browserName }) => {
-  it.skip(browserName === 'webkit', 'No third party cookies in WebKit');
-  it.skip(process.env.PW_CLOCK === 'frozen');
-  const context = await contextFactory({
-    ignoreHTTPSErrors: true,
-  });
-  const page = await context.newPage();
-
-  httpsServer.setRoute('/empty.html', (req, res) => {
-    res.writeHead(200, {
-      'Content-Type': 'text/html'
-    });
-    res.end(`<iframe src="${httpsServer.CROSS_PROCESS_PREFIX}/grid.html"></iframe>`);
-  });
-
-  httpsServer.setRoute('/grid.html', (req, res) => {
-    res.writeHead(200, {
-      'Set-Cookie': ['a=b; Path=/; Max-Age=3600; SameSite=None; Secure'],
-      'Content-Type': 'text/html'
-    });
-    res.end(`Hello world
-    <script>
-    setTimeout(() => fetch('/json'), 1000);
-    </script>`);
-  });
-
-  const cookie = new Promise(f => {
-    httpsServer.setRoute('/json', (req, res) => {
-      f(req.headers.cookie);
-      res.end();
-    });
-  });
-
-  await page.goto(httpsServer.EMPTY_PAGE);
-  expect(await cookie).toBe('a=b');
-});
-
 it('should allow unnamed cookies', async ({ page, context, server, browserName, platform }) => {
   server.setRoute('/cookies', (req, res) => {
     res.write(req.headers.cookie ?? 'undefined-on-server');
